+++ api/sync15/blobstorage.go
@@ -1,9 +1,8 @@
 package sync15
 
 import (
-	"bytes"
+	"fmt"
 	"io"
-	"net/http"
 
 	"github.com/juruen/rmapi/config"
 	"github.com/juruen/rmapi/log"
@@ -22,106 +21,51 @@ func NewBlobStorage(http *transport.HttpClientCtx) *BlobStorage {
 	}
 }
 
-const ROOT_NAME = "root"
-
-func (b *BlobStorage) PutRootUrl(hash string, gen int64) (string, int64, error) {
-	log.Trace.Println("fetching  ROOT url for: " + hash)
-	req := model.BlobRootStorageRequest{
-		Method:       http.MethodPut,
-		RelativePath: ROOT_NAME,
-		RootSchema:   hash,
-		Generation:   gen,
-	}
-	var res model.BlobStorageResponse
-
-	if err := b.http.Post(transport.UserBearer, config.UploadBlob, req, &res); err != nil {
-		return "", 0, err
-	}
-	return res.Url, res.MaxUploadSizeBytes, nil
-}
-func (b *BlobStorage) PutUrl(hash string) (string, int64, error) {
-	log.Trace.Println("fetching PUT blob url for: " + hash)
-	var req model.BlobStorageRequest
-	var res model.BlobStorageResponse
-	req.Method = http.MethodPut
-	req.RelativePath = hash
-	if err := b.http.Post(transport.UserBearer, config.UploadBlob, req, &res); err != nil {
-		return "", 0, err
-	}
-	return res.Url, res.MaxUploadSizeBytes, nil
+func (b *BlobStorage) GetReader(hash, filename string) (io.ReadCloser, error) {
+	return b.http.GetStream(transport.UserBearer, config.BlobUrl+hash, filename)
 }
 
-func (b *BlobStorage) GetUrl(hash string) (string, error) {
-	log.Trace.Println("fetching GET blob url for: " + hash)
-	var req model.BlobStorageRequest
-	var res model.BlobStorageResponse
-	req.Method = http.MethodGet
-	req.RelativePath = hash
-	if err := b.http.Post(transport.UserBearer, config.DownloadBlob, req, &res); err != nil {
-		return "", err
-	}
-	return res.Url, nil
+func (b *BlobStorage) UploadBlob(hash, filename string, reader io.Reader) error {
+	log.Trace.Println("uploading blob ", filename)
+
+	return b.http.PutStream(transport.UserBearer, config.BlobUrl+hash, reader, filename)
 }
 
-func (b *BlobStorage) GetReader(hash string) (io.ReadCloser, error) {
-	url, err := b.GetUrl(hash)
-	if err != nil {
-		return nil, err
-	}
-	log.Trace.Println("get url: " + url)
-
-	blob, _, err := b.http.GetBlobStream(url)
-	return blob, err
-}
-
-func (b *BlobStorage) UploadBlob(hash string, reader io.Reader) error {
-	url, size, err := b.PutUrl(hash)
-	if err != nil {
-		return err
-	}
-	log.Trace.Println("put url: " + url)
-
-	return b.http.PutBlobStream(url, reader, size)
-}
-
-// SyncComplete notifies that the sync is done
+// SyncComplete no longer used
 func (b *BlobStorage) SyncComplete(gen int64) error {
-	req := model.SyncCompletedRequest{
+	return nil
+}
+
+func (b *BlobStorage) WriteRootIndex(roothash string, gen int64, notify bool) (int64, error) {
+	log.Info.Println("writing root with gen: ", gen)
+	req := model.BlobRootStorageRequest{
+		Broadcast:  notify,
+		Hash:       roothash,
 		Generation: gen,
 	}
-	return b.http.Post(transport.UserBearer, config.SyncComplete, req, nil)
-}
+	var res model.BlobRootStorageResponse
+	headers := map[string]string{
+		transport.RmFileNameHeader: "roothash",
+	}
 
-func (b *BlobStorage) WriteRootIndex(roothash string, gen int64) (int64, error) {
-	log.Info.Println("writing root with gen: ", gen)
-	url, maxRequestSize, err := b.PutRootUrl(roothash, gen)
+	err := b.http.Put(transport.UserBearer, config.RootPut, req, &res, headers)
 	if err != nil {
 		return 0, err
 	}
-	log.Trace.Println("got root url:", url)
-	reader := bytes.NewBufferString(roothash)
+	if res.Hash != roothash {
+		return 0, fmt.Errorf("bug? root hash mismatch")
+	}
 
-	return b.http.PutRootBlobStream(url, gen, maxRequestSize, reader)
+	return res.Generation, nil
 }
 func (b *BlobStorage) GetRootIndex() (string, int64, error) {
-	url, err := b.GetUrl(ROOT_NAME)
+	var res model.BlobRootStorageResponse
+	err := b.http.Get(transport.UserBearer, config.RootGet, nil, &res)
 	if err != nil {
 		return "", 0, err
 	}
-	log.Info.Println("got root get url:", url)
-	blob, gen, err := b.http.GetBlobStream(url)
-	if err == transport.ErrNotFound {
-		return "", 0, nil
 
-	}
-	if err != nil {
-		return "", 0, err
-	}
-	content, err := io.ReadAll(blob)
-	if err != nil {
-		return "", 0, err
-	}
-	log.Info.Println("got root gen:", gen)
-	return string(content), gen, nil
+	log.Info.Println("got root gen:", res.Generation)
+	return res.Hash, res.Generation, nil
 
 }