+++ filetree/filetree.go
@@ -21,16 +21,26 @@ func (ctx *FileTreeCtx) Clear() {
 	ctx.root.Children = nil
 }
 
+const TrashID = "trash"
+
 func CreateFileTreeCtx() FileTreeCtx {
 	root := model.CreateNode(model.Document{
 		ID:          "",
 		Type:        "CollectionType",
 		VisibleName: "/",
 	})
+	trash := model.CreateNode(model.Document{
+		ID:           TrashID,
+		Type:         "CollectionType",
+		VisibleName: "trash",
+	})
+	root.Children[TrashID] = &trash
 
 	return FileTreeCtx{
 		&root,
-		make(map[string]*model.Node),
+		map[string]*model.Node{
+			TrashID: &trash,
+		},
 		make(map[string]map[string]struct{}),
 	}
 }
@@ -51,6 +61,16 @@ func (ctx *FileTreeCtx) NodeById(id string) *model.Node {
 	}
 }
 
+// FinishAdd add all nodes with missing parents to root
+func (ctx *FileTreeCtx) FinishAdd() {
+	for parentId, pendingChildren := range ctx.pendingParent {
+		for childId := range pendingChildren {
+			ctx.idToNode[childId].Parent = ctx.root
+			ctx.root.Children[childId] = ctx.idToNode[childId]
+		}
+		delete(ctx.pendingParent, parentId)
+	}
+}
 func (ctx *FileTreeCtx) AddDocument(document *model.Document) {
 	node := model.CreateNode(*document)
 	nodeId := document.ID
@@ -108,6 +128,86 @@ func (ctx *FileTreeCtx) MoveNode(src, dst *model.Node) {
 	}
 }
 
+// NodesByPath returns multiple nodes that match a pattern
+//
+// # use cases
+// dirname*		if 1 match list children, list dir names
+// dirname		list children
+// dirname/		list children
+// dirname/*	list children
+func (ctx *FileTreeCtx) NodesByPath(path string, currentNode *model.Node, ignoreTrailingSlash bool) ([]*model.Node, error) {
+	if currentNode == nil {
+		currentNode = ctx.Root()
+	}
+
+	entries := util.SplitPath(path)
+	length := len(entries)
+
+	if length == 0 {
+		return []*model.Node{currentNode}, nil
+	}
+
+	i := 0
+	if entries[i] == "" {
+		currentNode = ctx.Root()
+		i++
+	}
+
+	var err error
+	var result []*model.Node
+
+	for i < length {
+		isLast := i == length-1
+		entry := entries[i]
+
+		if entry == "" || entry == "." {
+			i++
+			continue
+		}
+
+		if entry == ".." {
+			if currentNode.Parent == nil {
+				currentNode = ctx.Root()
+			} else {
+				currentNode = currentNode.Parent
+			}
+
+			i++
+			continue
+		}
+
+		if isLast {
+			result, err = currentNode.FindByPattern(entry)
+		} else {
+			currentNode, err = currentNode.FindByName(entry)
+		}
+
+		if err != nil {
+			return nil, err
+		}
+
+		i++
+	}
+	switch len(result) {
+	case 0:
+		//handles dir/ in mv (rename vs move into)
+		if currentNode.IsDirectory() && ignoreTrailingSlash {
+			return currentNode.Nodes(), nil
+		}
+
+		return []*model.Node{currentNode}, nil
+	case 1:
+		//handles: dir/ , dir
+		if result[0].IsDirectory() && ignoreTrailingSlash {
+			return result[0].Nodes(), nil
+		}
+		//1 or more results
+		fallthrough
+	default:
+		return result, nil
+	}
+
+}
 func (ctx *FileTreeCtx) NodeByPath(path string, current *model.Node) (*model.Node, error) {
 	if current == nil {
 		current = ctx.Root()